requirements.txt
fastapi
uvicorn
openai
pdfplumber
faiss-cpu
pydantic
requests
python-dotenv
